# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "账本管理系统的后端 API 接口文档",
        "title": "账本系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "your-email@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5200",
    "basePath": "/",
    "paths": {
        "/classification": {
            "get": {
                "description": "获取所有支出分类，按树形结构返回",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "支出分类"
                ],
                "summary": "获取支出分类树",
                "responses": {
                    "200": {
                        "description": "成功返回分类树",
                        "schema": {
                            "$ref": "#/definitions/model.Response-array_service_TreeNodeDTO"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/model.Response-any"
                        }
                    }
                }
            },
            "put": {
                "description": "替换数据库中的所有支出分类数据，前端传入完整的树结构",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "支出分类"
                ],
                "summary": "更新整个支出分类树",
                "parameters": [
                    {
                        "description": "分类树数据",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.UpdateTreeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/model.Response-bool"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/model.Response-any"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/model.Response-any"
                        }
                    }
                }
            }
        },
        "/income-classification": {
            "get": {
                "description": "获取所有收入分类，按树形结构返回",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "收入分类"
                ],
                "summary": "获取收入分类树",
                "responses": {
                    "200": {
                        "description": "成功返回分类树",
                        "schema": {
                            "$ref": "#/definitions/model.Response-array_service_TreeNodeDTO"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/model.Response-any"
                        }
                    }
                }
            },
            "put": {
                "description": "替换数据库中的所有收入分类数据，前端传入完整的树结构",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "收入分类"
                ],
                "summary": "更新整个收入分类树",
                "parameters": [
                    {
                        "description": "分类树数据",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.UpdateTreeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/model.Response-bool"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/model.Response-any"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/model.Response-any"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Response-any": {
            "type": "object",
            "required": [
                "code",
                "data",
                "msg",
                "success"
            ],
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "model.Response-array_service_TreeNodeDTO": {
            "type": "object",
            "required": [
                "code",
                "data",
                "msg",
                "success"
            ],
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.TreeNodeDTO"
                    }
                },
                "msg": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "model.Response-bool": {
            "type": "object",
            "required": [
                "code",
                "data",
                "msg",
                "success"
            ],
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "boolean"
                },
                "msg": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "service.TreeNodeDTO": {
            "type": "object",
            "required": [
                "id",
                "name"
            ],
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.TreeNodeDTO"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "service.UpdateTreeRequest": {
            "type": "object",
            "required": [
                "tree"
            ],
            "properties": {
                "tree": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.TreeNodeDTO"
                    }
                }
            }
        }
    }
}